Load env vars at runtime instead of compiling them into our server.js: EZM-988 (#763)
This is a new PR following feedback given on https://github.com/ezcater/ezcater-ezmanage-ui/pull/682

This updates our env vars to be loaded at runtime instead of being compiled in at compile time. This is necessary because on the skyline pipeline, we're building in a docker instance that has different environment variables than where the code will eventually be run.

Note that this removes the `RAZZLE_` prefix from these vars so we don't accidentally try to use them the old way, but then have it not work as expected. This means people are going to have to update their .env files, and we'll have to update heroku (staging, demo, and prod) before this can get merged in.

This works by adding `window.__ENV__` to our html.js, which then is used in client.js to extract the vars from the server:
```
...

    <script charSet="utf-8">window.__ENV__={"braintreeKey":"sandbox_vdxbkzgx_4qn6grx84yjwvxfz","buildTarget":"server","ezrailsDomain":"http://localhost:3002","graphqlHost":"http://localhost:3002","intercomAppId":"","isProduction":false,"isServer":false,"isTest":false,"sentryEnvironment":"development"}</script>

...
```

Before merging this in I'll test this on staging. Assuming all goes well there, I'll send out an email, ping people in slack, update heroku, then merge in in.