Alternative way to handle margin with ezalerts (#1192)
After speaking with @ayip8 we've decided to hold off on any direct manipulation of recipe components in favor of applying margin to all children of content sections. This may seem weird and roundabout way to tackle things, but it does remove a potentially abused feature which is passing of `className` down. That would get things like emotion's `styled` to work, but at the cost of people abusing the `className` property. The styles look identical, but with no direct manipulation of recipe components.