Switch back to using react modal for EzModal (FAJITA-549)
FAJITA-549

Reverts the commit to swap from ReactModal to ReachUI.

In the process of bumping ezManage to use the latest version of recipe, we noticed that users had to double-click buttons in order for the button click to be triggered. We've narrowed down the problem to Semantic UI components inside ReachUI (which is the package we are currently using to build our modals). We know that the ReactModal library will handle this gracefully.