fix: playground padding clips content
The playground control was applying the padding outside of the iframe, causing anything with an outline to be clipped, rather than overlapping the padding.

This change moves the padding inside the frame, and adds some smarts to adjust the padding based on the content size, such that the if the content fills the playground, no additional padding is applied.