fix FOUC in environments that support emotion SSR
when rendering global styles, we have to wait for JS to be parsed and executed until the global styles are injected. This can lead to a flash of unstyled content (fouc).

In environments configured for emotion's server-side rendering (SSR), styles are collected and inserted into the head of the page during server render, removing the dependency on a JS runtime for the initial render.

Use emotion's `injectGlobal` method (scoped to a container) to provide global styles in order to support SSR, while still maintaining ability to scope style to a frame and also remove styles on unmount.