Set line height for EzInlineFeedback (#233)
Updated EzInlineFeedback's padding rule with a new line-height rule.

The EzInlineFeedback element is rendered when a status prop
is provided to EzToggle. Prior to this change, the appearance or
disappearance of the EzInlineFeedback element would cause the
EzToggle to visually shift a couple of pixels. Making this change
prevents that unwanted movement.

This fixes issue #223.