use `active` for activeClassName in EzNavigation links
This is a stepping stone for the upgrade to Emotion 10. In Emotion 10,
the `css` helper from `@emotion/core` returns an object where the `css`
helper from `emotion` and `react-emotion` both return a className string.

To get ahead on this change, we will no longer generate a className for
the active class via the `css` helper, and instead will use the string
'active'.

This has the added benefit of allowing us to access the theme for
providing the background-color.