add an onMouseDown to prevent a weird focus event being fired in internet explorer (#254)
Adds a mouseDown handler on the options in EzSelect to prevent default event propagation. We are doing this to fix a bug in internet explorer where after mouseDown a focusin event is fired on the body. This focusin is being picked up by our handlers which looks if focus has left the select box and closes the box. With this solution we now prevent firing a focus event when one of the options is clicked on.

I confirmed this change works in ie11, edge, chrome, safari (mobile and desktop), and firefox.

Why are we doing this?
Fix for #252