use emotion 10 `as` prop (avoid withComponent)
Emotion 10 supports the `as` prop on all it's styled components, so we
no longer have to implement this feature ourselves.

https://emotion.sh/docs/styled#as-prop

NOTE: This change was necessary for IE support too, as it seems the
`withComponent` feature of emotion causes multiple "target" properties
to be defined for the component, triggering the following error in IE:
"multiple definitions of a property not allowed in strict mode".