use emotion styled to compose multiple styles
This is a stepping stone for the upgrade to Emotion 10.

In Emotion 10,
the `css` helper from `@emotion/core` returns an object where the `css`
helper from `emotion` and `react-emotion` both return a className string. Accessing the className for a component/style is no longer easily available in Emotion 10.

To get ahead on this change, we will no longer generate a className for
the `alignBaseline` style to compose into the SVG icon style, and instead will use the styled helper to extend base components with new styles.