fix EzField[type=select] regression
Both button and inputs serialize the `value` property as a string so they can be submitted as part of the form, however, since EzField has previously allowed values of type String, Number and Boolean, we introduced a regression.

As a workaround, we'll publish the change event off of the container element, instead of the underlying focused element.