use label position for error messages for dropdown elements
This allows date/time/select (which all render a dropdown) to position their label consistenly.

While the original design was for the error icon to appear inside the input element, this location is already occupied by the chevron icon of the dropdown. Repositioning the error icon within the error icon would have a knock-on impact on the position of the error message flyout (the flyout pointer would need to be positioned conditionally), adding complexity.