conditionally apply error border collapse
In order to decouple the DOM hierarchy of the rendered input from the styles that apply the collapsed borders (error state), we need to apply the styles directly to the element, rather than via a descendant selector. Doing this affords us more flexibility in structuring EzField without the risk of breaking the styling.

using `css` prop allows us to apply this style without having to wrap the dynamic component with a new styled component at runtime.

fixes #302