hackery on playground resizing
It appears that `getComputedStyle(el).marginLeft` doesn't return the expect margin when some of our inner components are using absolute positioning... instead, we can calculate the applied automatic margin by subtracting the width of the element from the width of the parent.

Also fixed a bug with the existing resizing logic where components that are full-width ended up with an unwanted vertical margin, due to the short-circuiting logic for zero margins bailing out too early.