use @popperjs/core to position EzPopover
Fixes #313
Fixes #240

Note: Since we no longer control the positioning of the tooltip,
it no longer is easily possible to mock out the browser APIs in order to
be able to run the visual regression tests; we don't control which
browser APIs are used. That said, popperjs has test coverage for the
behavior that was previously under test, so we gain little duplicating
those tests here.