fix: EzLayout spacing reset
EzLayout uses the `responsive` helper for defining spacing values. The "split" variant doesn't have any spacing styles defined as it doesn't require any.

Fixing `responsive` such that the reset is still applied for the variant regardless, so that styles from the prior breakpoint are reset when the breakpoint changes to the new variant.

This code will stil bail if there is no variant for the specified breakpoint, such that the prior breakpoints styles still remain in play if they aren't overridden.