remove z-index causing icon to render above flyout
fixes #326

Without the z-index on the icon, the relative positioning of the input element and it's container caused the input/container elements to be positioned above the icon. This is because positioned elements are stacked in order of appearance in the HTML.

Since the z-index is a little heavy-handed (and causes the icon to also appear above the flyout), and we instead want to rely on the order of the HTML to determine stacking order, we can instead achieve the desired stacking order by removing the positioning styles of the input and container elements, thereby allowing the absolutely positioned icon to appear above them.