fix input focus bug
the css prop usage was causing the Input element to unmount/remount instead of re-rendering. It is unclear why this way happening - perhaps an emotion bug?

Removing the css prop usage an instead, pushing the relevant styles down onto EzTextInput (which is the underlying component that is used to style text inputs) fixes the issue.

I provided test coverage to catch to protect against future regressions.

Closes #328