EzButton tertiary (w icon) match height of other controls
Making all controls a consistent height makes aligning items a LOT easier, in particular for aligning items in a toolbar (which may have select controls, buttons and tertiary actions).

Ideally this change would apply to tertiary buttons as a whole, but since that would be a breaking change, let's hold off on that until we know the impact/benefits.