remove unused code
setSelectedOption exists purely to ensure a re-render when the selected option changes - the selectedOption value is unused.

Initializing activeOptionRef with selectedOption is also confusing, (and incorrect) since this ref is to hold the active option DOM element (not the underlying data). Since this value is always initially null (and doesn't update across renders), the incorrect code was a noop.