support peek of next/prev slide on carousel
To ensure that the next/prev buttons overlap the next/prev carousel item,
we need to center the items on the carousel; snapping items to the start
does not provision enough space for the next/prev buttons.

To use `scroll-snap-align: center`, we need to always have a middle
item, so we're inserting a pseudo after (i.e. between) every item and
applying the `scroll-snap-align: center` style to that element.