fix: incorrect page index calculated for last page
when we're on the last page, there may be fewer items than the page size,
and instead of being center justified, page is justified so that the
last item is always at the far right of the scroll container.

This means that the current index calculation needs to special case the
last page.