fix: move recipe-global import to top-level
The module that imports the recipe-globals in recipe's code base was the themes module. This is unfortunate, since ultimately we'll want to move away from importing this module at all. As it stands today, recipe doesn't work correctly if the theme isn't imported (and used sufficiently to avoid dead-code elimination in the bundler from omitting the import).