fix: calculation for itemWidth is over estimate
The first page and last page of items are sometimes wider than the other items (to account for the space otherwise provisioned for next/prev button). This means that in aggregate, the itemWidth calculations account too much space, causing other calculations (like number of items per page) to be incorrect.

Use the average item size instead to ensure calculations based on the scroll width are correct.