fix: item padding for incomplete pages isn't responsive
Since the number of items on a page can vary based on available space, the extra padding applied when the number of items is less than a complete page also needs to be applied conditionally.

This means that we need to "reset" the padding, if a larger screen size detects that padding is not required.