fix: rebalance carousel when there are only 2 pages
To ensure the carousel is centered when there aren't enough items to fill the final page, we apply what remains of the button-width padding the to end of the carousel.

This extra padding is not needed when there are only two pages in the carousel, since the items on the first and last page are evenly sized (it's only the "middle" pages of the carousel that are smaller).