derive storybook stories from component readme
The implementation here is definitely not the finished article; the mechanism i'm using for dynamically loading stories [storiesOf](src/stories/all.stories.tsx#L38) is deprecated in favor of the statically analyzable CSF format. But given that [Storybook doesn't current support dynamically loaded stories](https://github.com/storybookjs/storybook/issues/12654#issuecomment-711999247), this seemed like a pretty good place to start. The "recommended" alternative is building a webpack custom loader,  but that seemed like a task we could defer until we know this experiment is providing value.

When and if we get to that, the shopify team have already created a markdown loader that does EXACTLY what we need: https://github.com/Shopify/polaris-react/tree/master/.storybook.