fix: incorrectly sized carousel items
Today, the carousel resizes it's items to provision space for the next/previous buttons when the "peek" option is enabled. This resizing doesn't need to happen when there are fewer items that the page size (since we don't need to show the next/prev buttons).

That being said, we need to ensure that the CSS used to resize the carousel items is reset at larger breakpoints irrespective of whether the larger breakpoint needs to resize the items, to ensure than styles from smaller breakpoints don't interfere with the styles for larger breakpoints.

Closes #477