fix: timeline event example styling
Since the component in the `as` prop is not passing through the className to the rendered element, the EzLink styles are not applied.

We could pass through the class name, but as the `as` prop is not significant for this documented example, we can switch the example to use `href` instead of the `to`/`as` prop pair to keep things simple.