fix: auto-open in EzModal docs preventing dismiss
The auto-open behavior is nice as it allows folks can get an at-a-glance view of what the various modals look like, however when the modal is not mounted within an iframe, this logic can cause mutliple modals to appear and block dismissal.

Added some logic to short-circuit the auto-open logic if the modal is mounted into the doc-site outside of an iframe, and offer a button to open/close the modal.