update snitches to 0.0.10
The 0.0.10 release includes a fix to avoid duplicating media query
styles (over and over) for every component to guarantee the order of
media query based styles.

We no longer rely on the duplication strategy, as we can choose more
narrower media query ranges with (min|max)-width to avoid conflicts.