Enable Tree-shaking in consuming applications (#732)
## What did we change?

This adds `sideEffects` to the root package.json, informing webpack which files can be safely tree-shaken if left unused. It also updates the `@recipe-ui` packages whose sideEffects listed CSS files that no longer exist

## Why are we doing this?

Now that IE11 support has been dropped I am working on moderizing the store-next build a bit. We noticed awhile back that recipe was not correctly tree shaking unused components, but we chaulked it up to the transpiling of the recipe library that was confusing webpack.

Once I removed the transpiliation step I noticed that recipe was still including the full bundle. Then I realized that we didn't have `sideEffects` in the root package.json, which was causing webpack to include all of recipe regardless of the components being used.