workflow: fix release script so that it doesn't release before merging the version changes PR (#825)
## What did we change?

This removes the `yarn changeset version` and `bin/sync-package-versions` commands from the release script, and bumps the version of recipe to 12.0.5 so that the next release will be 12.0.6

## Why are we doing this?

When I updated the release script in #824 I did not realize calling `yarn changeset version` would break it. The github changeset action has magic built into it that determines it if should update the version PR or instead release to npm/github. It does this by checking for any version changes, which the newly introduced `yarn changeset version` does

Sadly when I merged the above PR it released 12.0.4 (`main`) as 12.0.5 but without any of the package version bumps, which will result in errors if consumers were to upgrade. I have since pulled that version, and as we've learned we cannot re-release a version so the only option is to skip it