Fix a memory leak and support using React 18's useId hook (if present) (#864)
## What did we change?

This addresses a memory leak we're seeing in our production apps that use Recipe. The `useUniqueId` method is holding onto the `setState` method returned from `useState` that is never cleared during an SSR render cycle, leading to an object that grows infinitely

This also updates the `useUniqueId` hook so that it uses the React 18 `useId` hook if present, falling back to our custom implementation when using React 17 or lower.

As part of this I had to remove the `mergeIds` method. I don't know what this might break, worse case it'll always take the ID from the last object passed to the function (mimicking what `Object.assign` would do)

## Why are we doing this?

* Memory leaks are bad
* In order for store to upgrade to react 18 we need to fix hydration errors, which are happening with any component using the custom `useUniqueId` hook (the IDs will always be mismatched between the SSR and hydration cycle)

## Screenshot(s) / Gif(s):

The memory leak we're seeing:

https://user-images.githubusercontent.com/1180841/192811565-8a2f2336-c54b-4117-92ce-fd2190984c70.png