Update EzRadio and EzRadioGroup props to match MUI (#882)
## What did we change?

This updates the typescript signature for the `onChange` callback for the `EzRadio` and `EzRadioGroup` to match MUI

## Why are we doing this?

I ran into this trying to pass an onChange method to `EzRadioGroup` when upgrading store-next. In order to get the value and checked status I need to do `event.target.value`/`event.target.checked`, but the typescript types in recipe doesn't expose an event so I get a typescript error. Functionally it works so it's an easy fix!

Ref that I used:

- [Radio props](https://mui.com/material-ui/api/radio/#props)
- [RadioGroup props](https://mui.com/material-ui/api/radio-group/#props)

For good measure I checked the `EzCheckbox` props and they match what MUI uses so I think we're good there!